rename file to blocklist
