+++ functions/api/image/nfts/asset/[[index]].tsx
@@ -2,7 +2,7 @@
 import { ImageResponse } from '@vercel/og'
 import React from 'react'
 
-import { blocklistedCollections } from '../../../../../src/nft/utils/constants'
+import { blocklistedCollections } from '../../../../../src/nft/utils/blocklist'
 import { WATERMARK_URL } from '../../../../constants'
 import getAsset from '../../../../utils/getAsset'
 import getFont from '../../../../utils/getFont'
