+++ src/nft/utils/index.ts
@@ -1,9 +1,9 @@
 export * from './asset'
+export * from './blocklist'
 export * from './buildActivityAsset'
 export * from './buildSellObject'
 export * from './carousel'
 export * from './collection'
-export * from './constants'
 export * from './currency'
 export * from './formatEventProperties'
 export * from './isAudio'
